backend/requirements.txt
yfinance==0.2.18
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.24.3
matplotlib==3.8.0
scikit-learn==1.3.2
transformers==4.36.0
torch==2.1.1
tokenizers==0.15.0
accelerate==0.25.0
bitsandbytes==0.41.3
psutil==5.9.6

# FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# Database dependencies
psycopg2-binary==2.9.7
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4 